fix : MarketGem PriceSet할때 int로 형변환

+++ Assets/Scripts/Market/MarketGem.cs
@@ -20,7 +20,7 @@ public class MarketGem : MonoBehaviour
     }
     public void PriceSet()
     {
-        curPrice = curGem? Random.Range(curGem.minPrice, curGem.maxPrice) : curPrice;
+        curPrice = curGem? (int)Random.Range(curGem.minPrice, curGem.maxPrice) : curPrice;
         price.text = curGem? curPrice.ToString() + "G" : price.text;
     }
